Added neighbor scanning and ship buying

+++ src/main/java/cz/cuni/mff/pijalekj/managers/CriminalsManager.java
@@ -4,16 +4,16 @@ import cz.cuni.mff.pijalekj.constants.Constants;
 
 import java.util.HashMap;
 
-public class CriminalsManager extends HashMap<Long, Integer> {
-    public boolean isCriminal(long entityID) {
+public class CriminalsManager extends HashMap<Integer, Integer> {
+    public boolean isCriminal(int entityID) {
         return this.containsKey(entityID);
     }
 
-    public void addCriminal(long entityID) {
+    public void addCriminal(int entityID) {
         this.put(entityID, Constants.DEFAULT_TIMEOUT);
     }
 
-    public void removeCriminal(long entityID) {
+    public void removeCriminal(int entityID) {
         this.remove(entityID);
     }
